fix: move out the UserRepositoryInterface into User directory
