+++ app/Repositories/User/UserRepositoryInterface.php
@@ -1,11 +1,10 @@
 <?php
 
-namespace App\Repositories\User\Eloquent;
+namespace App\Repositories\User;
 
-use App\Models\User;
 use App\Repositories\BaseRepositoryInterface;
 
 interface UserRepositoryInterface extends BaseRepositoryInterface
 {
-    public function getByEmail($email): ?User;
+    public function getByEmail($email): ?object;
 }
